girar-task-add: filter out unwanted specsubst tags
The situation in /gears repository when more than one specsubst tag
points to the same commit object is not desirable because it causes
problems for "task rebuild" operation.

To avoid this unwanted situation, when a build from a specsubst tag
is requested, remove all other specsubst tags from the subtask
repository.