gb: reimplement force-rebuild
When an arch build fails, gb stores failed subtask id of this arch and
then after the task build fails it chooses the least of stored failed
subtask ids of all the archs. On the next task run gb forces rebuild
all the packages of subtasks id equal or greater that chosen id.

This implementation is more robust than previous one: it is immune to
subtask deletion.