synchronize gb-toplevel-build with gb-toplevel-commit finishing its current work
While gb-toplevel-commit "is busy" (is looking for/sees some work to do),
it holds an exclusive lock on $GB_TASKS_DIR/index/$GB_REPO_NAME/pending.
This can be used for other processes to wait. If it has no more work,
it releases the lock and waits for new pending tasks.

(The rolling-in of the loop into another inner loop is basically an
equivalent rewriting. The only slight difference in behavior I see is
that if it detected a global cutout switch after "being busy", it used
to exit immediately, but now it would wait a while for inotifywait
after exiting from the inner loop subprocess before exiting from the
main process.)

gb-toplevel-build now waits while gb-toplevel-commit "is busy"
before it decides what to do.