gb-toplevel-commit: let tasks with greater iter be processed further before committing other ones
Give gb-toplevel-build a chance to start building tasks
with iter >= $min_iter, before we recalculate min_iter.

We might have just put such a task into awaiting state, and if we
continue straight away, we might worsen the situation for it by
committing a task with a lesser iter after recalculating min_iter.

This is more fair in the sense that no task should need too many
iterations before it can be committed. (At the cost of not committing
tasks with few iterations early.)

However, after some thought, I think that it's not obviously a clever
thing to do, because if we have put a pending task (with a big
iteration) into awaiting state, that still means that it would need to
re-check the build environment, and committing another task doesn't
necessarily worsen the amount of work needed for it. Especially, since
we have introduced the lock between the committer and the builder, there is
expected less useless work when rebuilding that task.

I suggest not to apply this patch yet; rather to observe and discuss.