gb-task-check-build-i: check_sid: relax sid check
If the sid check passes for the native repository because the specified
sid was found in the src index of the native repository, skip the check
for other repositories.

We have to relax the check this way because src indices for all
repositories except sisyphus are currently broken in several ways.