gb-commit-task: exit faster to release the lock in gb-toplevel-commit
gb-y-awake-tasks is not dependent on the task data, so it can be done
in the background independently of the other final actions.

The other actions first use the task data, then cleanup it, then move
it; so, it looks safer to do them sequentially. But I also suggest to
do this in the background.