girar-socket-forward-repo: bail out if the latest has changed
When the script is called, $repo/release is locked, so
$repo/release/latest cannot change.

When the script enters the archiving stage, the peer is released
and will subsequently unlock $repo/release, allowing for the latest
to be changed.

If the archiving stage takes a lot of time, there is a chance that the
latest will change and the new latest will be ready for archiving.

Avoid archiving the new latest by comparing it with the original one
and bailing out if the latest has changed.