girar-task-run: fix the race between postponed and completed tasks (ALT#37561)
When girar-task-run places the task into POSTPONED queue, there is a
chance that the status of dependencies listed in task/depends might have
changed since the last check while the task was not in POSTPONED state.

In that case, the task will remain in POSTPONED state forever because
its task/depends could not have been updated since the last check and
nothing is going to perform this update unless girar-task-run does it.

Suggested-by: Ivan Zakharyaschev <imz@altlinux.org>
Resolves: https://bugzilla.altlinux.org/show_bug.cgi?id=37561