girar-task-run: skip nevr check for pure rebuild subtasks
Firstly, evr of rebuild subtasks are permitted to change as result
of rebuild, so applying the evr check for them is not 100% correct.

Secondly, sids of pure rebuild subtasks have already been verified
by check_rebuilds, and that check can be relied upon.

There is a little chance that the package that's going to be rebuilt
has already been updated and removed from the target repository,
but the nevr check doesn't handle this case anyway.