bin/girar-task-approve: workaround users disconnections
It seems that certain users are using Ctrl+C instead of Ctrl+D to
finalize stdin input when crafting the approval message.  While certain
terminal emulators, like xterm(1), disregard the Ctrl+C in this
scenario, others, such as qterminal(1), proceed to terminate the ssh(1)
client process.  Unfortunately, the script on the girar side remains
oblivious to the termination of the ssh connection, persisting until it
tries to write something to stderr or stdout, resulting in failure
due to the inability to write to the closed terminal.

The approach of this fix involves writing an escape sequence to the
terminal right after obtaining the user message.  In cases where the
user has disconnected, this script will gracefully terminate to prevent
any alteration of the task.

Reported-by: Danil Shein <dshein@altlinux.org>